resolved retrofit error

+++ Pilot/app/src/main/java/com/amzur/pilot/pojos/ItemPojo.java
@@ -24,7 +24,7 @@ public class ItemPojo {
 
     public String imageUrl;
 
-    public int serialNumber;
+    public long serialNumber;
 
     public String itemCondition;
 }
